story-gs > staging: refactored steps and exports
