+++ src/bin/games/genius_square/features/constants.rs
@@ -1,0 +1,11 @@
+/// ----------------------------------------------------------------
+/// IMPORTS
+/// ----------------------------------------------------------------
+
+use std::time::Duration;
+
+/// ----------------------------------------------------------------
+/// CONSTANTS
+/// ----------------------------------------------------------------
+
+pub const TIMEOUT: Duration = Duration::from_secs(10);
