+++ src/bin/games/genius_square/features/mod.rs
@@ -1,3 +1,8 @@
 /// Highest logic of application.
 
-pub mod setup_game;
+pub mod constants;
+pub mod feat_initialise_game;
+pub mod feat_solve_game;
+
+pub use feat_initialise_game::feature_initialise_game;
+pub use feat_solve_game::feature_solve_game;
